refactor(plugin-registry): factor PluginArticle hooks into ../../hooks
Splits the inline `useCatalogEntry`, `useInstalledVersionTag`,
`useVersionPicker`, and `usePluginActions` hooks out of
`PluginArticle.tsx` into individual files under `src/hooks/`. Also
renames the existing hook files from kebab-case to camelCase
(`use-foo.ts` → `useFoo.ts`) and updates the barrel re-exports.
