+++ packages/plugins/plugin-code/src/capabilities/app-graph-builder.ts
@@ -28,14 +28,11 @@ import { makePluginSpecSubject } from '../plugin-spec';
 // package in this monorepo. The connector below maps a plugin's
 // `Plugin.Meta.spec` (relative to its package root) back to one of these
 // entries by looking for a key that ends with `/plugin-<dir>/<spec>`.
-const PLUGIN_MDLS = import.meta.glob<string>(
-  ['../../../plugin-*/PLUGIN.mdl', '../../../plugin-*/**/PLUGIN.mdl'],
-  {
-    query: '?raw',
-    import: 'default',
-    eager: true,
-  },
-);
+const PLUGIN_MDLS = import.meta.glob<string>(['../../../plugin-*/PLUGIN.mdl', '../../../plugin-*/**/PLUGIN.mdl'], {
+  query: '?raw',
+  import: 'default',
+  eager: true,
+});
 
 const resolveSpecContent = (pluginId: string, specPath: string): string | undefined => {
   const dirSegment = pluginId.replace(/^org\.dxos\.plugin\./, '');
