+++ packages/plugins/plugin-registry/src/hooks/index.ts
@@ -2,7 +2,11 @@
 // Copyright 2026 DXOS.org
 //
 
-export * from './use-auto-tags';
-export * from './use-disable-confirmation';
-export * from './use-registry-plugins';
-export * from './use-update-available';
+export * from './useAutoTags';
+export * from './useCatalogEntry';
+export * from './useDisableConfirmation';
+export * from './useInstalledVersionTag';
+export * from './usePluginActions';
+export * from './useRegistryPlugins';
+export * from './useUpdateAvailable';
+export * from './useVersionPicker';
