+++ packages/plugins/plugin-registry/src/hooks/useAutoTags.ts
