+++ packages/plugins/plugin-registry/src/hooks/useDisableConfirmation.ts
