+++ packages/plugins/plugin-registry/src/hooks/useRegistryPlugins.ts
