+++ packages/plugins/plugin-registry/src/hooks/useUpdateAvailable.ts
